backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
networkx
google-generativeai
requests